Move cloud flow message to correct folder
